+++ src/main/java/br/com/banco/service/TransferenciaService.java
@@ -20,7 +20,7 @@ public class TransferenciaService{
         return repository.findAll();
     }
 
-    public Transferencia findById(Long id) {
+    public Transferencia findById(Integer id) {
         Optional<Transferencia> obj=repository.findById(id);
         return obj.get();
     }
